requirements.txt
bert_score == 0.3.9
elasticsearch == 7.13.0
faiss_gpu == 1.7.1.post2
Flask == 2.0.1
h5py == 3.2.1
ijson == 3.1.4
ipdb == 0.13.8
jieba == 0.42.1
joblib == 1.0.1
nlg_eval == 2.3
nltk == 3.6.2
numpy == 1.19.2
scipy == 1.6.3
torch == 1.8.1+cu101
tqdm == 4.61.0
transformers == 4.6.1
faiss == 1.5.3
scikit_learn == 1.0
PyYAML == 5.4.1
